+++ inc/html.class.php
@@ -1096,9 +1096,6 @@ HTML;
          Html::requireJs('notifications_ajax');
       }
 
-      $tpl_vars['css_files'][] = 'public/lib/jstree.css';
-      Html::requireJs('jstree');
-
       $tpl_vars['css_files'][] = 'public/lib/leaflet.css';
       Html::requireJs('leaflet');
 
@@ -1190,11 +1187,6 @@ HTML;
 
          if (in_array('clipboard', $jslibs)) {
             Html::requireJs('clipboard');
-         }
-
-         if (in_array('jstree', $jslibs)) {
-            $tpl_vars['css_files'][] = 'public/lib/jstree.css';
-            Html::requireJs('jstree');
          }
 
          if (in_array('charts', $jslibs)) {
@@ -6004,9 +5996,6 @@ JAVASCRIPT;
                }
             }
             break;
-         case 'jstree':
-            $_SESSION['glpi_js_toload'][$name][] = 'public/lib/jstree.js';
-            break;
          case 'gantt':
             $_SESSION['glpi_js_toload'][$name][] = 'public/lib/dhtmlx-gantt.js';
             $_SESSION['glpi_js_toload'][$name][] = 'js/gantt-helper.js';
@@ -6051,9 +6040,6 @@ JAVASCRIPT;
          case 'rack':
             $_SESSION['glpi_js_toload'][$name][] = 'js/rack.js';
             $_SESSION['glpi_js_toload'][$name][] = 'lib/jqueryplugins/jquery.ui.touch-punch.js';
-            break;
-         case 'jstree':
-            $_SESSION['glpi_js_toload'][$name][] = 'public/lib/jstree.js';
             break;
          case 'leaflet':
             $_SESSION['glpi_js_toload'][$name][] = 'public/lib/leaflet.js';